git/housekeeping: Handle repacking inside of the strategy switch
We're about to introduce a new repacking strategy that will not use
git-repack(1), but git-pack-objects(1). This will become kind of awkward
as we precompute options inside of the repack-strategy switch and then
execute the command outside of the switch, and that will be confusing
when there is one switch case that is more special than the others
because it would execute the logic and exit early already.

Refactor the code so that all switch cases perform the repack and exit
directly.