remote: Stop relying on an empty default branch
GetDefaultBranch returns nothing when there are no branches, but HEAD
will still have an unborn branch set as the default branch. This means
when we go to switch the default branch heuristic to only use HEAD we
will end up with different behaviour. So instead of relying on the
gitaly default branch heuristic, here we explicitly detect if the
default branch exists or not before checking default branch ancestry.