repository: Refactor SearchFilesByName tests for failed requests
The testcases for calls to SearchFilesByName with failed requests set up
the server manually and then verify that the requests fail as expected.
This has the downside that we're not in fact testing the RPC calls as we
instead invoke the functions directly. Furthermore, it also keeps us
from testing calls which are proxied by Praefect.

Inline these tests into the general table-driven test of this RPC so
that we reuse the same server infrastructure.