repository: Allow query with leading dash in SearchFilesByName
It is not currently possible to specify a query in SearchFilesByName
that has a leading dash. This is not by design though, but due to a bug
where we pass the query as a normal argument instead of as an argument
that gets passed after the `--` separator. The Git command factory will
thus rightfully complain that it is not allowed to pass arguments with a
leading dash.

Fix this by instead passing the query as a post-separator argument.

Changelog: fixed