repository: Allow query with leading dash in SearchFilesByContent
It is not currently possible to specify a query in SearchFilesByContent
that has a leading dash. This is not by design though, but due to a bug
in the way we assemble the Git command: while we pass `-e` to indicate
that the next argument is the query, we treat it as a simple flag
instead of a value-flag and continue to pass the actual query as a
normal argument. Consequentially, our Git command factory will then
rightfully complain that arguments must not start with a dash and cause
the request to fail.

Fix this by instead passing the query as a value flag.

Changelog: fixed