repository: Don't use hardcoded SHA1 object hash in SearchFilesByName
In order to parse trees generated by git-ls-tree(1) we need to know
about the length of the object hash. Callers thus pass that object hash
to our tree parser, but it is currently hard-coded to the SHA1 object
hash in the SearchFilesByName RPC.

In practice, this does not matter as we only use the object hash when
calling `NextEntry()` on the parser. But as we call `NextEntryPath()`,
which causes us to ignore the object hash and only return the path,
the RPC is in fact already compatible with SHA256. It is still confusing
though to see the hardcoded SHA1 object format, and it's arguably the
wrong thing to do.

Detect the object hash of the repository at hand at runtime and pass it
to the tree parser. Enable testing of both the SearchFilesByName and the
SearchFilesByContent RPCs with SHA256.