git/housekeeping: Write full-repack timestamp before computing options
When performing a full repack we want to update the full-repack
timestamp regardless of whether the repack is successful or not so that
we don't busy-spin around full repacks. The code flow is a bit confusing
though given that we compute options that we want git-repack(1) to run
with, then update the timestamp, and finally run the command itself.

Refactor the logic so that computing the options and running the command
is as close together as possible. This helps readability a bit and is a
step towards refactoring the repack-strategy-switch to come last in this
function.