objectpool: Verify creating empty object pools works as expected
At several times I noticed that creating clones of an empty repository
will cause us to end up with a SHA1-based repository even in the case
where the source repository uses SHA256.

Add a testcase for creating object pools from an empty source repository
to assert that the object format is set up as expected. As this seems to
work as expected it indicates that the issue only exists for a subset of
transports in Git.