housekeeping: Add logic to inhibit running git-pack-refs(1)
With the upcoming WAL (via TransactionManager), we need a way to inhibit
running and cancel ongoing calls to git-pack-refs(1), while the
TransactionManager is modifying refs.

This is because when git-pack-refs(1) is in progress, git would add
locks to the refs which could block the TransactionManager from
modifying the said refs. In short, the TransactionManager needs
exclusivity to the refs for modification.

To facilitate this, `packRefsIfNeeded` now only runs if there are no
inhibitors added. This is provided via the `repositoryStates` within the
`RepositoryManager`. The `repositoryStates` provides methods which allows
users to add inhibitors for `git-pack-refs(1)` too.

The provided methods are internal to the package and can be easily
exposed via the RepositoryManager when required.