git: Unconditionally ignore gitconfig files
A while ago we have introduced the `ignore_gitconfig` configuration. If
set, we will override GIT_CONFIG_SYSTEM and GIT_CONFIG_GLOBAL as well as
override XDG_CONFIG_HOME so that Git won't pick up gitconfig files found
in any of these scopes. The goal of this is that we only ever use the
Git configuration that is found either in Gitaly's `config.toml` or in
the repository-local gitconfig.

This toggle has been enabled in all distributions unconditionally
already and was scheduled for removal in v16.0. So let's remove that
toggle and unconditionally ignore any global- or system-level gitconfig
files.

Changelog: removed