Move default branch verification to processTransaction
Transaction's default branch update is currently being verified
in verifyReference. If the verification passes, the branch update
is set in the returned log entry. We're soon going to change
verifyReferences to not return a log entry but just the reference
updates. This makes it easier to compose the log entry from various
parts of the transaction. To facilitate this, move default branch
verification to happen directly in processTransaction.