Centralize transaction quarantining logic into a staging repository
The transaction's quarantine directory is currently being applied to
the repository in multiple places. This commit stores a quarantined
repository in the transaction that is used by all of the staging and
verification logic. This centralizes the logic in a single place, and
makes it easier to plug more logic needed for example when the repository
doesn't exist yet.