protoregistry: Fix segmentation fault with unset `oneOf` fields
We're extracting the repository field of a Protobuf message via generic
reflection. This is quite involved and requires a bunch of custom logic
that is not well-tested and hard to understand. As the code itself
rightfully points out: "Warning: this reflection filled function is full
of forbidden dark elf magic". And it doesn't under-deliver on that
promise.

Despite being "dark elf magic", the code is also buggy and can lead to
`nil` pointer accesses when parsing messages that have an unset `oneOf`
field that contains the field with the repository extension.

Refactor the code to use our new helper that uses Protobuf-specific
reflection to find fields with repository extensions. This simplifies
the codebase and fixes the mentioned `nil` pointer access.

It is somewhat expected that the new code doesn't perform as fast as the
old reflection-based code where we precomputed offsets. Benchmarks show
that this is a bit of a mixed bag though.

    # generic-reflection-based code
    BenchmarkMethodInfo/unset_target_repository-12         	  500785	      2226 ns/op
    BenchmarkMethodInfo/target_repository-12               	  521468	      2156 ns/op
    BenchmarkMethodInfo/target_object_pool-12              	  227521	      5159 ns/op

    # Protobuf-reflection-based code
    BenchmarkMethodInfo/unset_target_repository-12         	  854041	      1574 ns/op
    BenchmarkMethodInfo/target_repository-12               	  370221	      4762 ns/op
    BenchmarkMethodInfo/target_object_pool-12              	  189813	      5305 ns/op

As you can see, we're faster in the case where the repository is unset,
slower for resolving target repositories, and roughly the same when
resolving object pools. But given that we're in a networking context
here where we talk about milliseconds of latency, and given that we only
ever resolve repositories in the very first request for each RPC call,
this performance hit really should be fine. The improved maintainability
of the new code oughtweighs the cost.

Changelog: fixed