protoregistry: Refactor storage fields to use Protobuf reflection
We use the same kind of "dark elf magic" to access storage fields for
storage-scoped requests as we did to access repositories in repo-scoped
requests. Let's also refactor access to these storage fields to use the
new Protobuf-reflection based infrastructure so that both share the same
foundations. This will allow us to get rid of the old generic-reflection
logic altogether.