protoregistry: Split registry and method info into different files
Both implementation and tests of the registry and its method infos are
kind of intermingled with no clear separation of concerns regarding the
code layout. Let's properly split them up into separate implementation
and tests for each of the types.

While at it, start testing the method info structure with the SHA256
object format.