protoregistry: Allow `additional_repository`s to be unset
When fields are tagged with the `additional_repository` extension then
we will always try to extract those fields and rewrite them in Praefect.
If this fails, we return an error to the caller and reject the request.

This is needlessly rigid though: the coordinator really only needs info
about the target repository so that it can properly route the request to
the correct Gitaly node. The case where additional repositories are not
set shouldn't necessarily cause us to fail the request. In fact, there
may be (and indeed are) perfectly valid usecases where you want to only
set a repository field conditionally depending on certain parameters.

Refactor the code to allow the case where additional repositories are
not set. We will simply forward these requests to the corresponding
Gitaly node, which can then verify by itself whether it is okay for the
repository to not be set.

Changelog: changed