localrepo: Return OID when conflicts found in MergeTree
The `MergeTree` function executes `git-merge-tree(1)`, parses the output
and returns the information to the user. Currently, it exclusively
returns either the OID of the tree created (success path) or the parsed
error (failure path).

But `git-merge-tree(1)` provides the OID of the tree even during the
failure path. This is necessary information which is needed to find the
contents of the conflicted files. Let's also parse the OID during the
failure path and return it to the client.

This fix exposes/removes the hard-coding of SHA1 OIDs in the
`TestParseResult` tests.