conflicts: Extract out the Git2Go implementation
The Git2Go implementation of finding conflict files is currently in-lined
into `ListConflictFiles`. This makes it hard to add a feature flag to
switch to the upcoming git-merge-tree(1) implementation of the same. To
make things easier we extract out this code into a new function
`conflictFilesWithGit2Go`.

We will introduce a similar function for the git-merge-tree(1)
implementation in the upcoming commits. Then we can use a feature flag
to switch between the two implementations and eventually remove the
Git2Go implementation and inline the newer git-merge-tree(1)
implementation.