localrepo: Extend MergeTreeError with ConflictingFileInfo
The MergeTreeError only holds the conflicting file name which is the
output of `git-merge-tree(1)` when using the `--name-only` option. When
using `git-merge-tree(1)` for `ListConflictFiles`, we'll need additional
information such as the OID, mode and stage of the conflicting files. By
adding ConflictingFileInfo, we make the struct extendable for additional
fields.