Drop storage from database keys
The database keys used by the TransactionManager are currently in the
form of <storage_name>:<relative_path>. Gitaly lacks a persistent identifier
for the repositories, and they are generally identified by the relative
path. For now, it's not really good either given the relative path can
be changed through RenameRepository. This will be fixed in the future
by removing RenameRepository, after which the relative path is stable.

The storage part has additional considerations. First of all, it's not
stable given the storage name could technically be changed. It's also
unnecessary. The storages are independent of each other, and can be for
example moved to a different node. They're also independent failure domains,
given the failure of a storage doesn't affect another storage. Given this,
each storage will have its own database instance once we integrate the
write-ahead logging in Gitaly. Given each TransactionManager works within
a single storage, the database is unique to a single storage, and a
repository's relative path uniquely identifies a repository within a storage,
the storage name is unnnecessary in the database keys. This commit
thus removes it from the keys.

The PartitionManager itself is not yet updated to handle separate databases
for the storages. This will be done in a follow up.