coordinator: Refactor tests for directing mutating RPC messages
Even though the tests for the stream director with mutating RPC calls is
theoretically table-driven, most of the variable bits are actually hard
coded into the test body. This makes it hard to extend the table to
cover more cases.

Refactor the tests so that the setup of repositories happens in a setup
function provided by the test. Move information about the RPC calls and
expectations into the testcases, too.