Makefile: Gracefully fail get Go package name
When Go is not installed, the command `go list` will fail. This command
is unconditionally called by the Makefile to determine the Gitaly Go
package. With this change the command will no longer print the error to
stderr, but return `unknown` so it's more obvious something was wrong
when the result was used elsewhere in the build.