proto: Remove deprecated ReduplicateRepository RPC
The ReduplicateRepository RPC reduplicates all objects borrowed from an
alternate repository so that it can be removed from object pools. This
mechanism is unsafe though as there is a race between reduplicating the
objects and concurrent writes to the repository.

We have thus deprecated the RPC a while ago for removal in Gitaly 16.0.
Callers are instead asked to use DisconnectGitAlternates, which both
makes sure to reduplicate all objects and remove the alternates file in
a single step.

So given that downstream callers have migrated to the new RPC, let's
follow our own deprecation notice and remove the RPC.

Changelog: removed