Makefile: Don't specify Go compatibility level in `go mod tidy`
We're explicitly setting the Go compatibility version to Go 1.17 when
executing `go mod tidy`. This isn't required though as the tool will
automatically pick the previous release to the current version that is
set in the `go.mod` file.

Remove the explicit compatibility version so that there are less
locations we need to keep track of when updating Go.