Makefile: Drop cleanup of legacy Go module files
Back in the days our build process left `go.mod` and `go.sum` files
lying around in the `_build` directory, which caused some problems when
we introduced embedded binaries. We have thus introduces a workaround in
5ff293bf6c (Remove legacy go.mod and go.sum files from _build directory,
2022-07-26) to clean up any such files if they exist.

This workaround is now more than half a year old, which means that we
can expect the cleanup to have been executed for all repositories that
still had those old files. So let's drop this workaround.