Ignore source port in pack-objects limiting
After deployment, it turns out that remote IP baggage may consist of
source port. Technically, this inclusion is not wrong. Apparently, the
source ports are usually ephemeral ports of clients. It always rotates.
Hence, it breaks the pack-objects limiting in such cases.

This commit is to sanitize the remote IP before applying the limiter.