repository: Sync Git alternates during replication
When a repository is replicated via the `ReplicateRepository()` RPC,
object pool relationships are not recreated on the target repository.
This change introduces Git altnernates file syncing between source and
target repositories when the `replicate_object_pool` field is enabled
for the RPC.

This operation is currently limited to linking the target repository to
its object pool on the target storage. If the required target object
pool does not exist, the target repository is not linked and replication
continues as normal.