protoregistry: Introduce function to find fields by extension
When Praefect receives the initial request for an RPC, it needs to
extract information about repositories and storages so that it can route
the request to the correct Gitaly nodes and rewrite repositories as
required. This is implemented by using custom logic that walks through
the message type via the generic `reflect` package. This logic is hard
to understand and has some buggy edge cases.

The Protobuf package already gives us all the tooling we need to dissect
Protobuf messages generically via the `protoreflect` and `protorange`
packages. So instead of using custom logic, we can just make the best
use of preexisting logic.

Introduce two new functions to find Protobuf fields via a given
extension and implement a set of tests to verify they work as expected.
The new functions are not yet used, but will gain users in subsequent
commits.