coordinator: Reject routing mutators with different repo storages
As with the preceding commit, we have the same issue when routing
mutators where the first message contains two repositories with
different storages. The result is that we may end up rewriting the
repositories to a repository on the wrong storage.

As routing mutators also uses `rewrittenRepositoryMessage()` the bug
should already be fixed by the preceding commit. But we'd still try to
connect to the wrong storage node first and will then try to resolve the
additional repository, which is likely to fail. The result would be a
confusing error message that the repository wasn't found, where the real
issue is that we got a request with differing repository storages.

Add an explicit check to catch this error condition so that we can
return a proper error message to the client.

Changelog: fixed