proto: Drop root tree ID from TreeEntry message
In 9eba19b5b (proto: Deprecate `root_oid` field in the TreeEntry
message, 2023-03-15) we have deprecated the `root_oid` field in the tree
entry message. This field was expensive to compute and requires us to
spawn a separate process. Given that there are no downstream users of
the root tree ID this is a complete waste of compute resources.

Remove the field.

Changelog: removed