catfile: Stop computing root tree ID
Now that we have dropped the `root_oid` field from the `TreeEntry`
message there is no need anymore to compute the root tree ID for tree
entries. Stop doing so to reduce the amount of work we need to do. This
should result in a speedup in all RPCs which use this, like the
`TreeEntry()`, `GetTreeEntries()` or `GetBlobs()`.