localrepo: Use `-z` flag for `git-merge-tree(1)`
Add the `-z` flag for `git-merge-tree(1)`. This makes the line separator
to be `\x00` and ends filenames with `\x00`. This makes it safer to
machine parse the output (for e.g. filenames with `\n`) and also adds
additional information in the 'Informational messages' section wherein
we can identify the paths associated with a particular conflict.

We also extend the test with some new scenarios:
1. single file rename conflict
2. single file conflict with ancestor
3. multi file conflict with deletion