Improve concurrency limiting logs
We employ multiple levels of concurrency limitation and it is crucial to
have observability to comprehend how requests are being limited.
Although Gitaly has some useful logs, particularly about queuing states,
they are not integrated with the gRPC logs, making it difficult to
understand the situation. To address this, we have unified all
concurrency-limiting logs in one location through this commit.
Furthermore, we have added new logging concurrent-limiting fields to
enhance the logging capabilities.