Move pack-objects logs to gRPC logs
Lately, the pack-objects RPC has been emitting logs that are difficult
to follow as they are each displayed on separate lines. This can be
problematic when attempting to investigate related issues. To address
this, a commit has been made to transfer these logs to gRPC logs. The
fields are stored in command.Stats, a FieldsProducer, and eventually
collated in a gRPC logging middleware.