praefect: Migrate remove-repository sub-command
Migration of the 'remove-repository' sub-command from the old
implementation to the new approach with usage of the
urfave/cli/v2 third party library. The invocation of the command
remains the same, but it has more descriptive representation when
help is invoked. Also, it is possible now to invoke help only for
the sub-command to see more details about it and its usage.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/5001