Refactor trace2.Manager to return trace
The trace2.Manager oversees the incorporation of trace2 and passes on
the resulting trace to managed hooks. Previously, the trace was not
accessible outside of this process, but with this update, the resulting
trace and any errors are returned. This trace can now be used to display
the trace in logs at a later time.