conflicts: Introduce and use `conflictFilesWithGitMergeTree`
The current implementation of `ListConflictFiles` uses Git2Go to find
the file conflicts, parses this data and returns the information to the
client.

As part of the plan to move away from Git2Go, we introduce a new
function `conflictFilesWithGitMergeTree` which uses git-merge-tree(1) to
find the conflicted files. We then use this implementation behind the
`ListConflictFilesMergeTree` featureflag.