Add limiter type to concurrency limiter
There are two versions of concurrency limiting: a generic per-rpc option
and a pack-objects option. These exist on different layers, with the
former being at the middleware layer and the latter located before
calling the git-pack-objects command. It can be difficult to distinguish
between the two from the logs. To address this, a limiter type has been
added to the concurrency monitor through this commit.