datastructure: Introduce new generic set type
The idiomatic way to implement sets in Go is to use maps that have an
empty structure as value. While this works alright, it can be hard to
read and doesn't provide a lot of niceties that sets can typically
provide.

With the recent addition of generics though the way generic data
structures can be implemented has been significantly improved. This also
changes what is idiomatic in favor of better encapsulated generic data
types that provide a higher-level abstraction level.

Introduce a new generic set type as a first internal implementation of
such data structures.