datastore: Rename generic `cacheValue` type
The cache for consistent storages uses a `cacheValue` type to represent
cached information. This name is too generic to be useful as it only
tells the reader that it has something to do with caching, but it does
not give any hint at what kind of information is actually getting
cached.

Rename the type to `cachedReplicaInfo` to further clarify its intent.