datastore: Refactor code to retrieve cached or fetch updated storages
The cache for consistent storages will only fetch consistent storages
when it got an invalidation notification. It thus got two modes: either
it can look up values from the cache, or it needs to fetch them anew.
This logic is split up across three different functions with some code
flow that is not exactly easy to follow as locking code is transgressing
across multiple calling contexts.

Refactor the code to instead be contained in a single function, only.
This makes it easier to follow and makes it more obvious how locking is
supposed to work.