datastore: Fix locking of consistent storages cache being too broad
When the lookup of cached consistent storages fails we need to lock the
cache in order to update it. The lock itself is supposed to only span
over a single key that consists of the relative path of the repository
we're about to look up so that concurrent callers can fetch consistent
storages for multiple repositories at once.

The lock is still too coarse though: while we indeed only create the
lock for the relative path, we lock on the level of the virtual
storages. In other words, if the same relative path exists on two
different virtual storages then the lock would noww span across both
virtual storages. This is needlessly broad and not really what the
code intends to do.

Fix this by refactoring the syncer to be per-virtual-storage.

Changelog: performance