Makefile: Upgrade Git to security release v2.40.1.gl2 and v2.39.3.gl1
It was decided on the git-security mailing list to release the "merge-tree:
load default git config" patch on the regular Git mailing list (which is
visible and open to everyone) as this is a security patch that most likely
affects only forges like GitLab, GitHub, etc, and not end users.

So the patch is not part of a Git security release and there is no specific
tag or branch for it, but despite this we need to handle it as an important
security fix that shouldn't be publicly disclosed before it is disclosed on
the main Git mailing list.

The date the patch will be disclosed on the Git mailing list has not yet been
decided, but it might be in the next few weeks.

Information about the patch is below.

```
merge-tree: load default git config

From 1f11b8550a22a0581b04dcd93654147af3ffe7ed Mon Sep 17 00:00:00 2001
From: Derrick Stolee <derrickstolee@github.com>
Date: Tue, 18 Apr 2023 14:30:57 -0400
Subject: [PATCH v2] merge-tree: load default git config

The 'git merge-tree' command handles creating root trees for merges
without using the worktree. This is a critical operation in many Git
hosts, as they typically store bare repositories.

This builtin does not load the default Git config, which can have
several important ramifications.

In particular, one config that is loaded by default is
core.useReplaceRefs. This is typically disabled in Git hosts due to
the ability to spoof commits in strange ways.

Since this config is not loaded specifically during merge-tree, users
were previously able to use refs/replace/ references to make pull
requests that looked valid but introduced malicious content. The
resulting merge commit would have the correct commit history, but the
malicious content would exist in the root tree of the merge.

The fix is simple: load the default Git config in cmd_merge_tree().
This may also fix other behaviors that are effected by reading default
config. The only possible downside is a little extra computation time
spent reading config. The config parsing is placed after basic argument
parsing so it does not slow down usage errors.

Helped-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Derrick Stolee <derrickstolee@github.com>
```

See the following MRs that merged that patch in our security Git repo:

  - https://gitlab.com/gitlab-org/security/git/-/merge_requests/18
  - https://gitlab.com/gitlab-org/security/git/-/merge_requests/19
  - https://gitlab.com/gitlab-org/security/git/-/merge_requests/20

Changelog: security