praefect/router: Split out assignment of repos to nodes
The RouteRepositoryCreation function handles many concerns at once:

    - It resolves the additional repository path.

    - It assigns the nodes on which the repository shall be created.

    - It derives the replica path that shall be used on the target
      nodes.

    - It reserves the repository ID of the new repository.

While most of the functionality is straight-forward, the logic to assign
a repository to its nodes is comparatively involved. This makes it hard
to figure out which parts belong together and ultimately makes the
function harder to extend.

Refactor the code by pulling out a function `assignRepositoryToNodes()`
that encapsulates the assignment logic. This makes the functions easier
to understand and extend.