praefect/router: Handle replication factor more explicitly
When routnig repository creations we have somewhat different logic based
on the replication factor:

    - If it is set to `1` then we only need to pick a primary node.

    - If it is set to `0` then we have to assign to all nodes.

    - If it is set to a value `n >= 2` then we need to pick `n - 1`
      secondaries.

The first case is handled via an explicit early return while the latter
two cases are mostly using the same codepath. The early return makes it
a bit harder to follow the logic though and also doesn't make it obvious
what input exactly is required for each of the different cases.

Furthermore, we're about to add another case that needs special
consideration. And while that case could also be handled via an early
return, it would only decrease readability even further.

Refactor the code to use an explicit switch over the different cases so
that it's easy to see which code paths belong together.